requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
dateparser==1.2.1
distro==1.9.0
faiss-cpu>=1.7.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community>=0.0.8
langchain-core==0.3.49
langchain-huggingface>=0.0.1
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.19
openai==1.69.0
orjson==3.10.16
packaging==24.2
pydantic==2.11.1
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sec-api==1.0.31
sentence-transformers>=2.2.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzlocal==5.3.1
urllib3==2.3.0
zstandard==0.23.0
